pyproject.toml
[project]
name = "tratativa-base-tratativa-solucao"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "attrs==25.4.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "fastparquet>=2026.3.0",
    "h11==0.16.0",
    "idna==3.11",
    "ntplib>=0.4.0",
    "openpyxl>=3.1.5",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pandas>=3.0.1",
    "psutil>=7.2.2",
    "pyarrow>=23.0.1",
    "pycparser==2.23",
    "pysocks==1.7.1",
    "python-dotenv==1.2.1",
    "pytz>=2026.1.post1",
    "requests==2.32.5",
    "selenium==4.38.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "trio==0.32.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.15.0",
    "urllib3==2.6.0",
    "webdriver-manager==4.0.2",
    "websocket-client==1.9.0",
    "wsproto==1.3.2",
]
